+++ camunda-bpm-assert/src/main/java/org/camunda/bpm/engine/test/assertions/cmmn/MilestoneAssert.java
@@ -1,4 +1,4 @@
-package org.camunda.bpm.engine.test.assertions.cmmn_new;
+package org.camunda.bpm.engine.test.assertions.cmmn;
 
 import org.camunda.bpm.engine.ProcessEngine;
