+++ camunda-bpm-assert/src/test/java/org/camunda/bpm/engine/test/assertions/cmmn/CmmnModelConstants.java
@@ -1,4 +1,4 @@
-package org.camunda.bpm.engine.test.assertions.cmmn_new;
+package org.camunda.bpm.engine.test.assertions.cmmn;
 
 /**
  * Constants for commonly used values.
