+++ camunda-bpm-assert/src/test/java/org/camunda/bpm/engine/test/assertions/cmmn/TaskAssertTest.java
@@ -1,4 +1,4 @@
-package org.camunda.bpm.engine.test.assertions.cmmn_new;
+package org.camunda.bpm.engine.test.assertions.cmmn;
 
 import org.camunda.bpm.engine.runtime.CaseInstance;
 import org.camunda.bpm.engine.test.Deployment;
@@ -10,7 +10,7 @@ import org.junit.runners.MethodSorters;
 
 import static org.camunda.bpm.engine.test.assertions.ProcessEngineAssertions.processEngine;
 import static org.camunda.bpm.engine.test.assertions.ProcessEngineTests.*;
-import static org.camunda.bpm.engine.test.assertions.cmmn_new.CmmnModelConstants.*;
+import static org.camunda.bpm.engine.test.assertions.cmmn.CmmnModelConstants.*;
 
 /**
  * @author Malte Sörensen <malte.soerensen@holisticon.de>
